fix(sdk/go): refuse chat-completions audio formats OpenRouter cannot deliver
OpenRouter's chat-completions audio modality only ever returns pcm16.
OpenAI rejects any other audio.format while streaming ("'audio.format'
does not support 'mp3' when stream=true"), and the OpenRouter gateway
rejects an audio completion that is not streamed at all ("Audio output
requires stream: true") — reproduced live against openai/gpt-audio and
openai/gpt-audio-mini with mp3, pcm16 and wav, and with
modalities:["audio"] alone. There is no request shape that returns
mp3/flac/opus from this route.

GenerateAudio now fails before the HTTP call when the requested format
is neither pcm16 nor wav, naming the requested format and the one the
provider actually delivers, instead of relaying a 400 that blames a
"stream" flag the caller never set. pcm16 and wav (the same samples
wrapped into a RIFF/WAVE container client-side) keep the streaming path
unchanged, and /audio/speech models such as hexgrad/kokoro-82m are
untouched — they still serve mp3.

This replaces the non-streaming JSON transport added earlier on this
branch (generateAudioViaChatJSON, longAudioClient, nonStreamAudioTimeout,
maxAudioResponseBytes and their tests): it turned one upstream 400 into a
different one and could never succeed against the real gateway.
